Updated users list and edit user form, added change password form

+++ app/main/user_routes.py
@@ -6,7 +6,7 @@ from app.models import db, User
 from app.main import bp
 #import datetime
 #from sqlalchemy import text
-from .user_forms import UserEditForm
+from .user_forms import UserEditForm, ChangePasswordForm
 
 
 @bp.route('/users')
@@ -15,23 +15,25 @@ def users():  # Main home page:
     if not current_user.is_admin():
       abort(403)
       
-    users = User.query.all()
-
+    users = User.query.order_by(User.display_name)
     return render_template('users.html',users = users)
 
 @bp.route('/user/new', methods=['GET', 'POST'],  defaults={'id':None})
-@bp.route('/user/<id>/edit', methods=['GET', 'POST'])
+@bp.route('/user/<int:id>/edit', methods=['GET', 'POST'])
 @login_required
 def edit_user(id):
+    print("id", id)
+    print("user", current_user.id)
     if not current_user.is_admin():
-      abort(403)
+      if id != current_user.id:
+        abort(403)
 
     if id:
       user = User.query.get_or_404(id)
       title = "Edit User"
     else:
       user = User()
-      title = "New User"
+      title = "New User"   
 
     form = UserEditForm(user=user)
     if form.validate_on_submit():
@@ -53,13 +55,14 @@ def edit_user(id):
         #if reset_password:
         #  flash(f"Password has been reset for user '{user.username}'")
 
-        return redirect(url_for('.users'))
+        return redirect(url_for('.show_user', id=user.id))
     elif request.method == 'GET':
+        form.id.data = user.id
         form.username.data = user.username
         form.display_name.data = user.display_name
         form.email.data = user.email
-        form.access.data = user.access
-    return render_template('user_edit.html', title=title, form=form)
+        form.access.data = user.access        
+    return render_template('user_edit.html', title=title, form=form, user=user)
 
 @bp.route('/user/<id>') 
 @login_required
@@ -67,13 +70,35 @@ def show_user(id):
    user = User.query.get_or_404(id)
    return render_template('show_user.html',user = user)
 
-@bp.route('/user/<id>/delete', methods=['POST'])
+@bp.route('/user/<int:id>/delete', methods=['POST'])
 @login_required
 def delete_user(id):
    if not current_user.is_admin():
       abort(403)
 
    user = User.query.get_or_404(id)
+
+   if user.id == current_user.id:   
+      flash('You cannot delete yourself', 'error')
+      return redirect(url_for('.show_user', id=user.id))          
+   
    db.session.delete(user)
    db.session.commit()
    return redirect(url_for('.users'))
+
+
+@bp.route('/user/change_password', methods=['GET', 'POST'])
+@login_required
+def change_password():
+    form = ChangePasswordForm()
+    if form.validate_on_submit():
+      if current_user.password_hash and not current_user.check_password(form.current_password.data):
+        flash('Your current password was invalid. Please try again','error')
+        return redirect(url_for('.change_password'))
+      if form.password.data:
+        first_password = not current_user.password_hash
+        current_user.set_password(form.password.data)
+        db.session.commit()
+        flash('Your password has been updated.')
+        return redirect(url_for('.home'))        
+    return render_template('user_change_password.html', title='Change Password', form=form)
